Fix untar and site install
Ansible sometimes has issues with lots of output on long running
processes causing Ansible to hang. This was the case with the
tar -xvzf operation. Simply removing -v resolved that problem.

Site install fails if -y is not specified to avoid interactive mode.
It will also fail if there is no way to send email on the server.
Finally, it will fail if it is not run from a clean state (DB and
settings.php exist)

Signed-off-by: FilmKnurd <andrew@astonishdesigns.com>